https://www.gofundme.com/f/preacher-six-film-release?utm_source=customer&utm_medium=copy_link&utm_campaign=p_cp+share-sheet
Let's get this film finished!! We are in a race to finish the film so we can have it out THIS FALL!!!. So here we are! 
We are in post production  and are film has a lot of effects for an indie film.  We need your help! 
We cannot do this without you. Jump on board!  No time like right now. This is only possible with your help! If you like original content that is driven by action and amazing writing you have come to the right place. No remakes, redoos , sequals, do overs for us! Lets take film back!! If you like Quentin Tarantino type films, you will love this one!! "Preacher Six" is like "Taxi Driver" meets "Sin City" meets "The Prophesy" Please be a part of something awesome!! 
      My name is Kyle Hester. Just a bit about me. I have been acting since I was in the 7th grade and this film thing is my life long passion. After acting for many years, I was given the opportunity to produce a film called "The Chair" I had the honor of working with many great actors including the late Roddy Piper in that film. Followed by "Zombie With a Shotgun" Do check it out.
       Well........... It is our turn. ￼My wife Tracy Ray is the screenwriter of "Preacher Six". She has created a world all of wonders that will leave you wanting to know more. I am truly blown away by the script personally! Preacher Six is a small town preacher that is summoned to the big city where he ends up fighting evil in a literal sense! The characters that he meets along the journey are something special and unique! They could all have their own films made about them! Grady Earls is our prolific director! You will see his graphics in the artwork. We are going for A list quality here! Speaking of high quality! Lets look at our cast!! Kyle Hester from "The Chair"
Peter Murnuk from "Justified" 
Carmen Argenziano from "Godfather II"!
Zach Galligan from "Gremlins"! 
Naomi Grossman from "American Horror Story"! 
Tomas Boykin from "Voodoo Possession"! 
David Banks from "Cut"! 
Lawrence Chau from Ghostly Encounters"!
Ezra Buzzington from "Fight Club"! 
Mary Le Gault from "Death Factory"! 
Jeremy Fultz from "Zombie With a Shotgun"! Plus many more!!  The only way all those actors  work will be seen  is with your help! Independent film only happens because of people like you! This Go Fund Me campaign pays for Everything from paying the cast and crew to the Insurance and film permits that make filming possible. The food that they eat on set is made possible because of you. Bottom line here is without you and your support all of this is a blank screen................... We are the new wave in Independent film and now....... So are you!!             
       Risks & Challenges What are the risks here? That no one contributes       anything and this film does not happen. That......... will NOT happen! Our community is strong and the need for Independent film is strong. We will keep going until this film is made. You know the expression "No plan B"? Well that is the case here! Come on board and be a part of something special! How else can you help? Please use the tools on this page to share the campaign on Twitter and your facebook! This is grass roots my friends and everything helps!! Thank you for taking the time to read this. Now...... Lets go make a film!!! ( All Perks regarding the completion of the film will be sent when the film is made. Ei. DVD, BluRay and signed poster from cast. 
$10 you get our everlasting gratitude! 
$25 you will get a signed postcard by the creators Kyle Hester and Tracy Ray 
$35 you get the above plus a streaming link to the film when it is ready. 
$55 you will get all the above plus a signed DVD when the film is done. 
$75 you get all the above plus a signed poster signed by the creators Kyle Hester and Tracy Ray. 
$100 all the above plus get your name in the credits as special thanks! 
$250 all the above plus you will be credited with FUNDING PRODUCER! In order to get your swag we will contact you for your information.
Send us the perk level and your mailing address so we cannot send you the swag when it is ready. We are the new wave in Independent film and now....... So are you!!
We dedicate this film to Carmen Argenziano. He was a great actor and greater man!
#RipCarmenArgenziano